requirements.txt
flask
openai>=1.56.0
fpdf2
python-dotenv
